UVMHS.Core.Data.Iter

Documentation

empty𝐼 :: 𝐼 a

cons𝐼 :: a -> 𝐼 a -> 𝐼 a

stream :: ToIter a t => t -> 𝑆 a

zipWith :: (ToIter a t₁, ToIter b t₂) => (a -> b -> c) -> t₁ -> t₂ -> 𝐼 c

zip :: (ToIter a t₁, ToIter b t₂) => t₁ -> t₂ -> 𝐼 (a ∧ b)

snoc𝐼 :: 𝐼 a -> a -> 𝐼 a

isEmpty :: ToIter a t => t -> 𝔹

firstElem :: ToIter a t => t -> 𝑂 a

append𝐼 :: 𝐼 a -> 𝐼 a -> 𝐼 a

mjoin𝐼 :: 𝐼 (𝐼 a) -> 𝐼 a

bind𝐼 :: 𝐼 a -> (a -> 𝐼 b) -> 𝐼 b

mapM𝐼 :: Monad m => (a -> m b) -> 𝐼 a -> m (𝐼 b)

fold :: ToIter a t => b -> (a -> b -> b) -> t -> b

foldFromWith :: ToIter a t => b -> (a -> b -> b) -> t -> b

foldFromOn :: ToIter a t => b -> t -> (a -> b -> b) -> b

foldOnFrom :: ToIter a t => t -> b -> (a -> b -> b) -> b

foldOnWith :: ToIter a t => t -> (a -> b -> b) -> b -> b

foldWithOn :: ToIter a t => (a -> b -> b) -> t -> b -> b

foldWithFrom :: ToIter a t => (a -> b -> b) -> b -> t -> b

foldk :: ToIter a t => b -> (a -> b -> (b -> b) -> b) -> t -> b

foldkFromWith :: ToIter a t => b -> (a -> b -> (b -> b) -> b) -> t -> b

foldkFromOn :: ToIter a t => b -> t -> (a -> b -> (b -> b) -> b) -> b

foldkOnFrom :: ToIter a t => t -> b -> (a -> b -> (b -> b) -> b) -> b

foldkOnWith :: ToIter a t => t -> (a -> b -> (b -> b) -> b) -> b -> b

foldkWithOn :: ToIter a t => (a -> b -> (b -> b) -> b) -> t -> b -> b

foldkWithFrom :: ToIter a t => (a -> b -> (b -> b) -> b) -> b -> t -> b

foldr :: ToIter a t => b -> (a -> b -> b) -> t -> b

foldrFromWith :: ToIter a t => b -> (a -> b -> b) -> t -> b

foldrFromOn :: ToIter a t => b -> t -> (a -> b -> b) -> b

foldrOnFrom :: ToIter a t => t -> b -> (a -> b -> b) -> b

foldrOnWith :: ToIter a t => t -> (a -> b -> b) -> b -> b

foldrWithOn :: ToIter a t => (a -> b -> b) -> t -> b -> b

foldrWithFrom :: ToIter a t => (a -> b -> b) -> b -> t -> b

mfold :: (Monad m, ToIter a t) => b -> (a -> b -> m b) -> t -> m b

mfoldFromWith :: (Monad m, ToIter a t) => b -> (a -> b -> m b) -> t -> m b

mfoldFromOn :: (Monad m, ToIter a t) => b -> t -> (a -> b -> m b) -> m b

mfoldOnFrom :: (Monad m, ToIter a t) => t -> b -> (a -> b -> m b) -> m b

mfoldOnWith :: (Monad m, ToIter a t) => t -> (a -> b -> m b) -> b -> m b

mfoldWithOn :: (Monad m, ToIter a t) => (a -> b -> m b) -> t -> b -> m b

mfoldWithFrom :: (Monad m, ToIter a t) => (a -> b -> m b) -> b -> t -> m b

mfoldk :: (Monad m, ToIter a t) => b -> (a -> b -> (m b -> m b) -> m b) -> t -> m b

mfoldkFromWith :: (Monad m, ToIter a t) => b -> (a -> b -> (m b -> m b) -> m b) -> t -> m b

mfoldkFromOn :: (Monad m, ToIter a t) => b -> t -> (a -> b -> (m b -> m b) -> m b) -> m b

mfoldkOnFrom :: (Monad m, ToIter a t) => t -> b -> (a -> b -> (m b -> m b) -> m b) -> m b

mfoldkOnWith :: (Monad m, ToIter a t) => t -> (a -> b -> (m b -> m b) -> m b) -> b -> m b

mfoldkWithOn :: (Monad m, ToIter a t) => (a -> b -> (m b -> m b) -> m b) -> t -> b -> m b

mfoldkWithFrom :: (Monad m, ToIter a t) => (a -> b -> (m b -> m b) -> m b) -> b -> t -> m b

mfoldr :: (Monad m, ToIter a t) => b -> (a -> b -> m b) -> t -> m b

mfoldrFromWith :: (Monad m, ToIter a t) => b -> (a -> b -> m b) -> t -> m b

mfoldrFromOn :: (Monad m, ToIter a t) => b -> t -> (a -> b -> m b) -> m b

mfoldrOnFrom :: (Monad m, ToIter a t) => t -> b -> (a -> b -> m b) -> m b

mfoldrOnWith :: (Monad m, ToIter a t) => t -> (a -> b -> m b) -> b -> m b

mfoldrWithOn :: (Monad m, ToIter a t) => (a -> b -> m b) -> t -> b -> m b

mfoldrWithFrom :: (Monad m, ToIter a t) => (a -> b -> m b) -> b -> t -> m b

eachWith :: (Monad m, ToIter a t) => (a -> m ()) -> t -> m ()

eachOn :: (Monad m, ToIter a t) => t -> (a -> m ()) -> m ()

eachkWith :: (Monad m, ToIter a t) => (a -> (m () -> m ()) -> m ()) -> t -> m ()

eachkOn :: (Monad m, ToIter a t) => t -> (a -> (m () -> m ()) -> m ()) -> m ()

exec :: (Monad m, ToIter (m ()) t) => t -> m ()

sum :: (ToIter a t, Additive a) => t -> a

product :: (ToIter a t, Multiplicative a) => t -> a

concat :: (Monoid a, ToIter a t) => t -> a

sequence :: (Seqoid a, ToIter a t) => t -> a

compose :: ToIter (a -> a) t => t -> a -> a

mcompose :: Monad m => ToIter (a -> m a) t => t -> a -> m a

wcompose :: Comonad w => ToIter (w a -> a) t => t -> w a -> a

minsFrom :: (ToIter a t, Ord a) => a -> t -> a

maxsFrom :: (ToIter a t, Ord a) => a -> t -> a

joinsFrom :: (ToIter a t, Join a) => a -> t -> a

joins :: (JoinLattice a, ToIter a t) => t -> a

meetsFrom :: (ToIter a t, Meet a) => a -> t -> a

meets :: (MeetLattice a, ToIter a t) => t -> a

or :: ToIter 𝔹 t => t -> 𝔹

orf :: ToIter (a -> 𝔹) t => t -> a -> 𝔹

andf :: ToIter (a -> 𝔹) t => t -> a -> 𝔹

and :: ToIter 𝔹 t => t -> 𝔹

count :: forall n t a. (Zero n, One n, Plus n, ToIter a t) => t -> n

countWith :: forall n t a. (Zero n, One n, Plus n, ToIter a t) => (a -> 𝔹) -> t -> n

reverse :: ToIter a t => t -> 𝐼 a

replicateI :: forall n a. (Eq n, Zero n, One n, Plus n) => n -> (n -> a) -> 𝐼 a

replicate :: forall n a. (Eq n, Zero n, One n, Plus n) => n -> a -> 𝐼 a

build :: forall n a. (Eq n, Zero n, One n, Plus n) => n -> a -> (a -> a) -> 𝐼 a

range :: (Eq n, Zero n, One n, Plus n, Minus n) => n -> n -> 𝐼 n

upto :: (Eq n, Zero n, One n, Plus n) => n -> 𝐼 n

reiter :: ToIter a t => s -> (a -> s -> s ∧ b) -> t -> 𝐼 b

withIndex :: forall n t a. (Zero n, One n, Plus n, ToIter a t) => t -> 𝐼 (n ∧ a)

withFirst :: ToIter a t => t -> 𝐼 (𝔹 ∧ a)

mapFirst :: ToIter a t => (a -> a) -> t -> 𝐼 a

mapAfterFirst :: ToIter a t => (a -> a) -> t -> 𝐼 a

keepN :: (ToIter a t, Eq n, Zero n, One n, Plus n) => n -> t -> 𝐼 a

withLast :: ToIter a t => t -> 𝐼 (𝔹 ∧ a)

mapLast :: ToIter a t => (a -> a) -> t -> 𝐼 a

mapLastOn :: ToIter a t => t -> (a -> a) -> 𝐼 a

mapBeforeLast :: ToIter a t => (a -> a) -> t -> 𝐼 a

filterMap :: ToIter a t => (a -> 𝑂 b) -> t -> 𝐼 b

filterMapOn :: ToIter a t => t -> (a -> 𝑂 b) -> 𝐼 b

filter :: ToIter a t => (a -> 𝔹) -> t -> 𝐼 a

filterOn :: ToIter a t => t -> (a -> 𝔹) -> 𝐼 a

inbetween :: ToIter a t => a -> t -> 𝐼 a

alignLeftFill :: ℂ -> ℕ -> 𝕊 -> 𝕊

alignLeft :: ℕ -> 𝕊 -> 𝕊

alignRightFill :: ℂ -> ℕ -> 𝕊 -> 𝕊

alignRight :: ℕ -> 𝕊 -> 𝕊

list :: ToIter a t => t -> 𝐿 a

lazyList :: ToIter a t => t -> [a]

string :: ToIter ℂ t => t -> 𝕊

stringS :: ToIter 𝕊 t => t -> 𝕊

truncate𝕊 :: ℕ64 -> 𝕊 -> 𝕊 -> 𝕊

showCollection :: ToIter a t => 𝕊 -> 𝕊 -> 𝕊 -> (a -> 𝕊) -> t -> 𝕊

showWith𝐼 :: (a -> 𝕊) -> 𝐼 a -> 𝕊

firstMaxByLT :: ToIter a t => (a -> a -> 𝔹) -> t -> 𝑂 a

sortWith :: ToIter a t => (a -> a -> Ordering) -> t -> 𝐿 a

sortOn :: (ToIter a t, Ord b) => (a -> b) -> t -> 𝐿 a

sort :: (ToIter a t, Ord a) => t -> 𝐿 a

materialize :: ToIter a t => t -> 𝐼 a

mapWhile :: (a -> a) -> (a -> 𝔹) -> 𝐼 a -> 𝐼 a

dropWhile :: (a -> 𝔹) -> 𝐼 a -> 𝐼 a

Orphan instances

All 𝔹

Instance details

Methods

all :: 𝐼 𝔹

All ()

Instance details

Methods

all :: 𝐼 ()

Functor 𝐼

Instance details

Methods

map :: (a -> b) -> 𝐼 a -> 𝐼 b

FunctorM 𝐼

Instance details

Methods

mapM :: Monad m => (a -> m b) -> 𝐼 a -> m (𝐼 b)

Bind 𝐼

Instance details

Methods

(≫=) :: 𝐼 a -> (a -> 𝐼 b) -> 𝐼 b

Monad 𝐼

Instance details

Return 𝐼

Instance details

Methods

return :: a -> 𝐼 a

Single a (𝐼 a)

Instance details

Methods

single :: a -> 𝐼 a

ToIter a (𝐼 a)

Instance details

Methods

iter :: 𝐼 a -> 𝐼 a

𝕊 ⇄ (𝐼 ℂ)

Instance details

Methods

isoto :: 𝕊 -> 𝐼 ℂ

isofr :: 𝐼 ℂ -> 𝕊

Show a => Show (𝐼 a)

Instance details

Methods

showsPrec :: Int -> 𝐼 a -> ShowS

show :: 𝐼 a -> String

showList :: [𝐼 a] -> ShowS

Show a => Show (𝐿 a)

Instance details

Methods

showsPrec :: Int -> 𝐿 a -> ShowS

show :: 𝐿 a -> String

showList :: [𝐿 a] -> ShowS

Show a => Show (𝑆 a)

Instance details

Methods

showsPrec :: Int -> 𝑆 a -> ShowS

show :: 𝑆 a -> String

showList :: [𝑆 a] -> ShowS

All a => All (𝑂 a)

Instance details

Methods

all :: 𝐼 (𝑂 a)

Append (𝐼 a)

Instance details

Methods

(⧺) :: 𝐼 a -> 𝐼 a -> 𝐼 a

Monoid (𝐼 a)

Instance details

Null (𝐼 a)

Instance details

Methods

null :: 𝐼 a

(All a, All b) => All (a ∧ b)

Instance details

Methods

all :: 𝐼 (a ∧ b)

(All a, All b) => All (a ∨ b)

Instance details

Methods

all :: 𝐼 (a ∨ b)