UVMHS.Core.IO

Documentation

type 𝑅 = IORef

ref :: a -> IO (𝑅 a)

deref :: 𝑅 a -> IO a

(↢) :: 𝑅 a -> a -> IO () infix 1

io_UNSAFE :: IO a -> a

initUVMHS :: IO ()

owrite :: 𝕊 -> IO ()

out :: 𝕊 -> IO ()

outs :: ToIter 𝕊 t => t -> IO ()

oflush :: IO ()

shout :: Show a => a -> IO ()

trace :: 𝕊 -> ()

traceM :: Monad m => 𝕊 -> m ()

ewrite :: 𝕊 -> IO ()

err :: 𝕊 -> IO ()

eflush :: IO ()

redirectErrToOut :: IO ()

iread :: IO 𝕊

iargs :: IO (𝐿 𝕊)

ilocalArgs :: 𝐿 𝕊 -> IO a -> IO a

abortIOCode :: ℤ64 -> IO a

abortIO :: IO a

exitIO :: IO a

failIO :: 𝕊 -> IO a

throwIO :: IOError -> IO a

catchIO :: IO a -> (IOError -> IO a) -> IO a

cleanExit :: IO a -> IO a

fread :: ℙ -> IO 𝕊

fwrite :: ℙ -> 𝕊 -> IO ()

fappend :: ℙ -> 𝕊 -> IO ()

fcopy :: ℙ -> ℙ -> IO ()

dfilesAll :: IO (𝐿 ℙ)

dfiles :: IO (𝐿 ℙ)

din :: ℙ -> IO a -> IO a

dtouch :: ℙ -> IO ()

drremove :: ℙ -> IO ()

dcurrent :: IO ℙ

pexists :: ℙ -> IO 𝔹

shell :: 𝕊 -> IO ((𝔹 ∧ 𝕊) ∧ 𝕊)

shellOK :: 𝕊 -> IO 𝕊

shelll :: 𝕊 -> IO ((𝔹 ∧ 𝕊) ∧ 𝕊)

shelllOK :: 𝕊 -> IO 𝕊

gc :: IO ()

time :: (() -> a) -> IO (a ∧ TimeD)

rtime :: 𝕊 -> (() -> a) -> IO a

timeIO :: IO a -> IO (a ∧ TimeD)

rtimeIO :: 𝕊 -> IO a -> IO a

profile :: IO a -> IO ((a ∧ 𝔻) ∧ 𝔻)

type IOError = IOException

The Haskell 2010 type for exceptions in the IO monad. Any I/O operation may raise an IOError instead of returning a result. For a more general type of exception, including also those that arise in pure code, see Exception.

In Haskell 2010, this is an opaque type.

data ExitCode

Defines the exit codes that a program can return.