UVMHS.Core.Init

Documentation

data Void

data Nat

Constructors

Z

S Nat

data P (a :: k)

Constructors

P

data Lazy a

Constructors

Lazy

Fields

unLazy :: ~a

type 𝔹 = Bool

newtype 𝐼 a

Constructors

𝐼

Fields

un𝐼 :: forall b. (a -> b -> (b -> b) -> b) -> b -> (b -> b) -> b

data 𝐿 a

Constructors

Nil

a :& (𝐿 a) infixr 8

data 𝑂 a

Constructors

None

Some a

newtype 𝑄 a

Constructors

𝑄

Fields

un𝑄 :: Seq a

type 𝕊 = Text

newtype k ⇰ v infixr 1

Constructors

𝐷

Fields

un𝐷 :: Map k v

data a ∨ b infixl 5

Constructors

Inl a

Inr b

data a ∧ b infixl 6

Constructors

a :* b infixl 8

type ℕ = Natural

type ℕ64 = Word64

type ℕ32 = Word32

type ℕ16 = Word16

type ℕ8 = Word8

type ℤ = Integer

type ℤ64 = Int64

type ℤ32 = Int32

type ℤ16 = Int16

type ℤ8 = Int8

type ℚ = Rational

type ℚᴾ = Ratio ℕ

type 𝔻 = Double

newtype 𝔻ᴾ

Constructors

𝔻ᴾ

Fields

un𝔻ᴾ :: 𝔻

data ℝ

Constructors

Integer ℤ

Rational ℚ

Double 𝔻

data ℝᴾ

Constructors

Natural ℕ

Rationalᴾ ℚᴾ

Doubleᴾ 𝔻ᴾ

type ℂ = Char

newtype 𝑆 a

Constructors

𝑆

Fields

un𝑆 :: () -> 𝑂 (a ∧ 𝑆 a)

newtype 𝑃 a

Constructors

𝑃

Fields

un𝑃 :: Set a

data (a :: k) ≟ (b :: k) :: ★ where

Constructors

Refl :: forall (k :: ★) (a :: k). a ≟ a

data W (c :: Constraint) where

Constructors

W :: c => W c

data Ex (t :: k -> ★) :: ★ where

Constructors

Ex :: forall (k :: ★) (t :: k -> ★) (a :: k). t a -> Ex t

data Ex_C (c :: k -> Constraint) (t :: k -> ★) :: ★ where

Constructors

Ex_C :: forall (k :: ★) (c :: k -> Constraint) (t :: k -> ★) (a :: k). c a => t a -> Ex_C c t

type STACK = HasCallStack

class CHS a b | b -> a where

Methods

tohs :: a -> b

frhs :: b -> a

fail :: forall (r :: RuntimeRep) (a :: TYPE r) m. STACK => [ℂ] -> m a

($) :: forall r a (b :: TYPE r). (a -> b) -> a -> b infixr 0

error :: forall (r :: RuntimeRep) (a :: TYPE r). STACK => 𝕊 -> a

uncurry :: ((a ∧ b) -> c) -> a -> b -> c

id :: a -> a

with :: W c -> (c => a) -> a

fromRational :: Rational -> 𝔻

negate :: ℕ -> ℤ

not :: 𝔹 -> 𝔹

const :: a -> b -> a

flip :: (a -> b -> c) -> b -> a -> c

rotateL :: (a -> b -> c -> d) -> b -> c -> a -> d

rotateR :: (a -> b -> c -> d) -> c -> a -> b -> d

curry :: (a -> b -> c) -> (a ∧ b) -> c

on :: (b -> b -> c) -> (a -> b) -> a -> a -> c

unpack :: forall (k :: ★) (t :: k -> ★) (b :: ★). Ex t -> (forall (a :: k). t a -> b) -> b

(⩔) :: 𝔹 -> 𝔹 -> 𝔹 infixl 2

(⩓) :: 𝔹 -> 𝔹 -> 𝔹 infixl 3

(∘) :: (b -> c) -> (a -> b) -> a -> c infixl 7

cond :: 𝔹 -> a -> a -> a

exfalso :: Void -> a

run𝐼 :: (b -> b) -> b -> (a -> b -> (b -> b) -> b) -> 𝐼 a -> b

run𝐼On :: 𝐼 a -> (b -> b) -> b -> (a -> b -> (b -> b) -> b) -> b

foldk𝐼 :: b -> (a -> b -> (b -> b) -> b) -> 𝐼 a -> b

fold𝐼 :: b -> (a -> b -> b) -> 𝐼 a -> b

foldr𝐼 :: b -> (a -> b -> b) -> 𝐼 a -> b

map𝐼 :: (a -> b) -> 𝐼 a -> 𝐼 b

null𝐼 :: 𝐼 a

single𝐼 :: a -> 𝐼 a

list𝐼 :: 𝐼 a -> 𝐿 a

iter𝐿 :: 𝐿 a -> 𝐼 a

lazyList𝐼 :: 𝐼 a -> [a]

iterLL :: [a] -> 𝐼 a

coerce_UNSAFE :: a -> b

weq_UNSAFE :: P a -> P b -> W (a ~ b)

void_UNSAFE :: Void

unpack_C :: forall (k :: ★) (c :: k -> Constraint) (t :: k -> ★) (b :: ★). Ex_C c t -> (forall (a :: k). c a => t a -> b) -> b

ifThenElse :: 𝔹 -> a -> a -> a

𝕤 :: [ℂ] -> 𝕊

𝕟 :: ℕ -> ℕ

𝕟64 :: ℕ -> ℕ64

𝕟32 :: ℕ -> ℕ32

𝕟16 :: ℕ -> ℕ16

𝕟8 :: ℕ -> ℕ8

𝕫 :: ℕ -> ℤ

𝕫64 :: ℕ -> ℤ64

𝕫32 :: ℕ -> ℤ32

𝕫16 :: ℕ -> ℤ16

𝕫8 :: ℕ -> ℤ8

𝕢 :: ℕ -> ℚ

𝕢ᴾ :: ℕ -> ℚᴾ

𝕕ᴾ :: 𝔻 -> 𝔻ᴾ

tohsChars :: 𝕊 -> [ℂ]

frhsChars :: [ℂ] -> 𝕊

appto :: a -> (a -> b) -> b

(∘∘) :: (c -> d) -> (a -> b -> c) -> a -> b -> d

(∘∘∘) :: (d -> e) -> (a -> b -> c -> d) -> a -> b -> c -> e

mirror :: (a -> b -> c -> d) -> c -> b -> a -> d

tohs𝑂F :: (a -> b) -> 𝑂 a -> Maybe b

tohs𝑂 :: 𝑂 a -> Maybe a

frhs𝑂F :: (a -> b) -> Maybe a -> 𝑂 b

frhs𝑂 :: Maybe a -> 𝑂 a

assert :: Bool -> a -> a

If the first argument evaluates to True, then the result is the second argument. Otherwise an AssertionFailed exception is raised, containing a String with the source file and line number of the call to assert.

Assertions can normally be turned on or off with a compiler flag (for GHC, assertions are normally on unless optimisation is turned on with -O or the -fignore-asserts option is given). When assertions are turned off, the first argument to assert is ignored, and the second argument is returned as the result.

class a ~R# b => Coercible (a :: k) (b :: k)

Coercible is a two-parameter class that has instances for types a and b if the compiler can infer that they have the same representation. This class does not have regular instances; instead they are created on-the-fly during type-checking. Trying to manually declare an instance of Coercible is an error.

Nevertheless one can pretend that the following three kinds of instances exist. First, as a trivial base-case:

instance Coercible a a

Furthermore, for every type constructor there is an instance that allows to coerce under the type constructor. For example, let D be a prototypical type constructor (data or newtype) with three type arguments, which have roles nominal, representational resp. phantom. Then there is an instance of the form

instance Coercible b b' => Coercible (D a b c) (D a b' c')

Note that the nominal type arguments are equal, the representational type arguments can differ, but need to have a Coercible instance themself, and the phantom type arguments can be changed arbitrarily.

The third kind of instance exists for every newtype NT = MkNT T and comes in two variants, namely

instance Coercible a T => Coercible a NT

instance Coercible T b => Coercible NT b

This instance is only usable if the constructor MkNT is in scope.

If, as a library author of a type constructor like Set a, you want to prevent a user of your module to write coerce :: Set T -> Set NT, you need to set the role of Set's type parameter to nominal, by writing

type role Set nominal

For more details about this feature, please refer to Safe Coercions by Joachim Breitner, Richard A. Eisenberg, Simon Peyton Jones and Stephanie Weirich.

Since: ghc-prim-0.4.0

coerce :: forall {k :: RuntimeRep} (a :: TYPE k) (b :: TYPE k). Coercible a b => a -> b

The function coerce allows you to safely convert between values of types that have the same representation with no run-time overhead. In the simplest case you can use it instead of a newtype constructor, to go from the newtype's concrete type to the abstract type. But it also works in more complicated settings, e.g. converting a list of newtypes to a list of concrete types.

This function is representation-polymorphic, but the RuntimeRep type argument is marked as Inferred, meaning that it is not available for visible type application. This means the typechecker will accept coerce Int Age 42.

type Constraint = CONSTRAINT LiftedRep

The kind of lifted constraints

data CallStack

CallStacks are a lightweight method of obtaining a partial call-stack at any point in the program.

A function can request its call-site with the HasCallStack constraint. For example, we can define

putStrLnWithCallStack :: HasCallStack => String -> IO ()

as a variant of putStrLn that will get its call-site and print it, along with the string given as argument. We can access the call-stack inside putStrLnWithCallStack with callStack.

>>> :{
putStrLnWithCallStack :: HasCallStack => String -> IO ()
putStrLnWithCallStack msg = do
  putStrLn msg
  putStrLn (prettyCallStack callStack)
:}

Thus, if we call putStrLnWithCallStack we will get a formatted call-stack alongside our string.

>>> putStrLnWithCallStack "hello"
hello
CallStack (from HasCallStack):
  putStrLnWithCallStack, called at <interactive>:... in interactive:Ghci...

GHC solves HasCallStack constraints in three steps:

If there is a CallStack in scope -- i.e. the enclosing function has a HasCallStack constraint -- GHC will append the new call-site to the existing CallStack.

If there is no CallStack in scope -- e.g. in the GHCi session above -- and the enclosing definition does not have an explicit type signature, GHC will infer a HasCallStack constraint for the enclosing definition (subject to the monomorphism restriction).

If there is no CallStack in scope and the enclosing definition has an explicit type signature, GHC will solve the HasCallStack constraint for the singleton CallStack containing just the current call-site.

CallStacks do not interact with the RTS and do not require compilation with -prof. On the other hand, as they are built up explicitly via the HasCallStack constraints, they will generally not contain as much information as the simulated call-stacks maintained by the RTS.

A CallStack is a [(String, SrcLoc)]. The String is the name of function that was called, the SrcLoc is the call-site. The list is ordered with the most recently called function at the head.

NOTE: The intrepid user may notice that HasCallStack is just an alias for an implicit parameter ?callStack :: CallStack. This is an implementation detail and should not be considered part of the CallStack API, we may decide to change the implementation in the future.

Since: base-4.8.1.0

callStack :: HasCallStack => CallStack

Return the current CallStack.

Does *not* include the call-site of callStack.

Since: base-4.9.0.0

withFrozenCallStack :: HasCallStack => (HasCallStack => a) -> a

Perform some computation without adding new entries to the CallStack.

Since: base-4.9.0.0

data Bool

Constructors

False

True

data Ordering

Constructors

LT

EQ

GT

data IO a

A value of type IO a is a computation which, when performed, does some I/O before returning a value of type a.

There is really only one way to "perform" an I/O action: bind it to Main.main in your program. When your program is run, the I/O will be performed. It isn't possible to perform I/O from an arbitrary function, unless that function is itself in the IO monad and called at some point, directly or indirectly, from Main.main.

IO is a monad, so IO actions can be combined using either the do-notation or the >> and >>= operations from the Monad class.

class Eq a where

The Eq class defines equality (==) and inequality (/=). All the basic datatypes exported by the Prelude are instances of Eq, and Eq may be derived for any datatype whose constituents are also instances of Eq.

The Haskell Report defines no laws for Eq. However, instances are encouraged to follow these properties:

Reflexivity

x == x = True

Symmetry

x == y = y == x

Transitivity

if x == y && y == z = True, then x == z = True

Extensionality

if x == y = True and f is a function whose return type is an instance of Eq, then f x == f y = True

Negation

x /= y = not (x == y)

Minimal complete definition: either == or /=.

Minimal complete definition

(==) | (/=)

Methods

(==) :: a -> a -> Bool infix 4

class Eq a => Ord a where

The Ord class is used for totally ordered datatypes.

Instances of Ord can be derived for any user-defined datatype whose constituent types are in Ord. The declared order of the constructors in the data declaration determines the ordering in derived Ord instances. The Ordering datatype allows a single comparison to determine the precise ordering of two objects.

Ord, as defined by the Haskell report, implements a total order and has the following properties:

Comparability

x <= y || y <= x = True

Transitivity

if x <= y && y <= z = True, then x <= z = True

Reflexivity

x <= x = True

Antisymmetry

if x <= y && y <= x = True, then x == y = True

The following operator interactions are expected to hold:

x >= y = y <= x

x < y = x <= y && x /= y

x > y = y < x

x < y = compare x y == LT

x > y = compare x y == GT

x == y = compare x y == EQ

min x y == if x <= y then x else y = True

max x y == if x >= y then x else y = True

Note that (7.) and (8.) do not require min and max to return either of their arguments. The result is merely required to equal one of the arguments in terms of (==).

Minimal complete definition: either compare or <=. Using compare can be more efficient for complex types.

Minimal complete definition

compare | (<=)

Methods

compare :: a -> a -> Ordering

class Show a where

Conversion of values to readable Strings.

Derived instances of Show have the following properties, which are compatible with derived instances of Read:

The result of show is a syntactically correct Haskell expression containing only constants, given the fixity declarations in force at the point where the type is declared. It contains only the constructor names defined in the data type, parentheses, and spaces. When labelled constructor fields are used, braces, commas, field names, and equal signs are also used.

If the constructor is defined to be an infix operator, then showsPrec will produce infix applications of the constructor.

the representation will be enclosed in parentheses if the precedence of the top-level constructor in x is less than d (associativity is ignored). Thus, if d is 0 then the result is never surrounded in parentheses; if d is 11 it is always surrounded in parentheses, unless it is an atomic expression.

If the constructor is defined using record syntax, then show will produce the record-syntax form, with the fields given in the same order as the original declaration.

For example, given the declarations

infixr 5 :^:
data Tree a =  Leaf a  |  Tree a :^: Tree a

the derived instance of Show is equivalent to

instance (Show a) => Show (Tree a) where

       showsPrec d (Leaf m) = showParen (d > app_prec) $
            showString "Leaf " . showsPrec (app_prec+1) m
         where app_prec = 10

       showsPrec d (u :^: v) = showParen (d > up_prec) $
            showsPrec (up_prec+1) u .
            showString " :^: "      .
            showsPrec (up_prec+1) v
         where up_prec = 5

Note that right-associativity of :^: is ignored. For example,

show (Leaf 1 :^: Leaf 2 :^: Leaf 3) produces the string "Leaf 1 :^: (Leaf 2 :^: Leaf 3)".

Minimal complete definition

showsPrec | show

Methods

show :: a -> String

A specialised variant of showsPrec, using precedence context zero, and returning an ordinary String.

otherwise :: Bool

otherwise is defined as the value True. It helps to make guards more readable. eg.

 f x | x < 0     = ...
     | otherwise = ...

fromInteger :: Num a => Integer -> a

Conversion from an Integer. An integer literal represents the application of the function fromInteger to the appropriate value of type Integer, so such literals have type (Num a) => a.

undefined :: forall (r :: RuntimeRep) (a :: TYPE r). HasCallStack => a

A special case of error. It is expected that compilers will recognize this and insert error messages which are more appropriate to the context in which undefined appears.

fromString :: IsString a => String -> a