UVMHS.Core.Vector

Documentation

newtype 𝕌 a

Constructors

𝕌

Fields

un𝕌 :: Vector a

newtype 𝕍Mut a

Constructors

𝕍Mut

Fields

un𝕍Mut :: IOVector a

newtype 𝕍 a

Constructors

𝕍

Fields

un𝕍 :: Vector a

vec :: ToIter a t => t -> 𝕍 a

vecC :: ToIterC a t => t -> 𝕍 a

vecF :: ℕ64 -> (ℕ64 -> a) -> 𝕍 a

vecDΩ :: (ℕ64 ⇰ a) -> 𝕍 a

iter𝕍 :: 𝕍 a -> 𝐼 a

show𝕍 :: Show a => 𝕍 a -> 𝕊

idx𝕍 :: ℕ64 -> 𝕍 a -> 𝑂 a

idx𝕍Ω :: ℕ64 -> 𝕍 a -> a

idx𝕍Ω_UNSAFE :: ℕ64 -> 𝕍 a -> a

map𝕍 :: (a -> b) -> 𝕍 a -> 𝕍 b

mapM𝕍 :: forall m a b. Monad m => (a -> m b) -> 𝕍 a -> m (𝕍 b)

null𝕍 :: Null a => ℕ64 -> 𝕍 a

csize𝕍 :: 𝕍 a -> ℕ64

vecIMut :: ToIter a t => t -> IO (𝕍Mut a)

vecVMut :: 𝕍 a -> IO (𝕍Mut a)

idx𝕍Mut :: ℕ64 -> 𝕍Mut a -> IO a

set𝕍Mut :: ℕ64 -> a -> 𝕍Mut a -> IO ()

eachI𝕍Mut :: (ℕ64 -> a -> IO ()) -> 𝕍Mut a -> IO ()

each𝕍Mut :: (a -> IO ()) -> 𝕍Mut a -> IO ()

values𝕍Mut :: 𝕍Mut a -> IO (𝕍 a)

grow𝕍Mut :: ℕ64 -> 𝕍Mut a -> IO (𝕍Mut a)

uvec :: (Storable a, ToIter a t) => t -> 𝕌 a

uvecC :: (Storable a, ToIterC a t) => t -> 𝕌 a

uvecF :: Storable a => ℕ64 -> (ℕ64 -> a) -> 𝕌 a

uvecDΩ :: Storable a => (ℕ64 ⇰ a) -> 𝕌 a

iter𝕌 :: Storable a => 𝕌 a -> 𝐼 a

show𝕌 :: (Storable a, Show a) => 𝕌 a -> 𝕊

idx𝕌 :: Storable a => ℕ64 -> 𝕌 a -> 𝑂 a

idx𝕌Ω :: Storable a => ℕ64 -> 𝕌 a -> a

idx𝕌Ω_UNSAFE :: Storable a => ℕ64 -> 𝕌 a -> a

map𝕌 :: (Storable a, Storable b) => (a -> b) -> 𝕌 a -> 𝕌 b

mapM𝕌 :: forall m a b. (Monad m, Storable a, Storable b) => (a -> m b) -> 𝕌 a -> m (𝕌 b)

null𝕌 :: (Storable a, Null a) => ℕ64 -> 𝕌 a

class Storable a where

The member functions of this class facilitate writing values of primitive types to raw memory (which may have been allocated with the above mentioned routines) and reading values from blocks of raw memory. The class, furthermore, includes support for computing the storage requirements and alignment restrictions of storable types.

Memory addresses are represented as values of type Ptr a, for some a which is an instance of class Storable. The type argument to Ptr helps provide some valuable type safety in FFI code (you can't mix pointers of different types without an explicit cast), while helping the Haskell type system figure out which marshalling method is needed for a given pointer.

All marshalling between Haskell and a foreign language ultimately boils down to translating Haskell data structures into the binary representation of a corresponding data structure of the foreign language and vice versa. To code this marshalling in Haskell, it is necessary to manipulate primitive data types stored in unstructured memory blocks. The class Storable facilitates this manipulation on all types for which it is instantiated, which are the standard basic types of Haskell, the fixed size Int types (Int8, Int16, Int32, Int64), the fixed size Word types (Word8, Word16, Word32, Word64), StablePtr, all types from Foreign.C.Types, as well as Ptr.

Minimal complete definition

sizeOf, alignment, (peek | peekElemOff | peekByteOff), (poke | pokeElemOff | pokeByteOff)

Methods

sizeOf :: a -> Int

Computes the storage requirements (in bytes) of the argument. The value of the argument is not used.

alignment :: a -> Int

Computes the alignment constraint of the argument. An alignment constraint x is fulfilled by any address divisible by x. The value of the argument is not used.

peekElemOff :: Ptr a -> Int -> IO a

Read a value from a memory area regarded as an array of values of the same kind. The first argument specifies the start address of the array and the second the index into the array (the first element of the array has index 0). The following equality holds,

peekElemOff addr idx = IOExts.fixIO $ \result ->
  peek (addr `plusPtr` (idx * sizeOf result))

Note that this is only a specification, not necessarily the concrete implementation of the function.

pokeElemOff :: Ptr a -> Int -> a -> IO ()

Write a value to a memory area regarded as an array of values of the same kind. The following equality holds:

pokeElemOff addr idx x = 
  poke (addr `plusPtr` (idx * sizeOf x)) x

peekByteOff :: Ptr b -> Int -> IO a

Read a value from a memory location given by a base address and offset. The following equality holds:

peekByteOff addr off = peek (addr `plusPtr` off)

pokeByteOff :: Ptr b -> Int -> a -> IO ()

Write a value to a memory location given by a base address and offset. The following equality holds:

pokeByteOff addr off x = poke (addr `plusPtr` off) x

peek :: Ptr a -> IO a

Read a value from the given memory location.

Note that the peek and poke functions might require properly aligned addresses to function correctly. This is architecture dependent; thus, portable code should ensure that when peeking or poking values of some type a, the alignment constraint for a, as given by the function alignment is fulfilled.

poke :: Ptr a -> a -> IO ()

Write the given value to the given memory location. Alignment restrictions might apply; see peek.