UVMHS.Init

Documentation

class CHS a b | b -> a where

Methods

tohs :: a -> b

frhs :: b -> a

data P (a :: k)

Constructors

P

data Nat

Constructors

Z

S Nat

data W (c :: Constraint) where

Constructors

W :: c => W c

newtype k ⇰ v infixr 2

Constructors

𝐷

Fields

un𝐷 :: Map k v

newtype 𝑃 a

Constructors

𝑃

Fields

un𝑃 :: Set a

newtype 𝑄 a

Constructors

𝑄

Fields

un𝑄 :: Seq a

data 𝑆 a where

Constructors

𝑆 :: forall s a. s -> (s -> 𝑂 (a ∧ s)) -> 𝑆 a

newtype 𝐼 a

Constructors

𝐼

Fields

un𝐼 :: forall b. (a -> b -> b) -> b -> b

data 𝐿 a

Constructors

Nil

a :& (𝐿 a) infixr 8

data 𝑂 a

Constructors

None

Some a

data a ∧ b infixl 5

Constructors

a :* b infixl 1

data a ∨ b infixl 4

Constructors

Inl a

Inr b

type 𝔹 = Bool

data Void

type 𝕊 = Text

type ℂ = Char

type 𝔻 = Double

type ℚ = Rational

type ℤ8 = Int8

type ℤ16 = Int16

type ℤ32 = Int32

type ℤ64 = Int64

type ℤ = Integer

type ℕ8 = Word8

type ℕ16 = Word16

type ℕ32 = Word32

type ℕ64 = Word64

type ℕ = Natural

fromString :: [ℂ] -> 𝕊

fromInteger :: ℤ -> ℕ

negate :: ℕ -> ℤ

fromRational :: Rational -> 𝔻

fail :: [ℂ] -> m a

𝕟64 :: ℕ -> ℕ64

𝕟32 :: ℕ -> ℕ32

𝕟16 :: ℕ -> ℕ16

𝕟8 :: ℕ -> ℕ8

𝕫 :: ℕ -> ℤ

𝕫64 :: ℕ -> ℤ64

𝕫32 :: ℕ -> ℤ32

𝕫16 :: ℕ -> ℤ16

𝕫8 :: ℕ -> ℤ8

chars :: 𝕊 -> [ℂ]

fromChars :: [ℂ] -> 𝕊

with :: W c -> (c => a) -> a

error :: 𝕊 -> a

id :: a -> a

const :: a -> b -> a

(∘) :: (b -> c) -> (a -> b) -> a -> c infixl 6

(∘∘) :: (c -> d) -> (a -> b -> c) -> a -> b -> d

flip :: (a -> b -> c) -> b -> a -> c

rotateR :: (a -> b -> c -> d) -> c -> a -> b -> d

rotateL :: (a -> b -> c -> d) -> b -> c -> a -> d

mirror :: (a -> b -> c -> d) -> c -> b -> a -> d

on :: (b -> b -> c) -> (a -> b) -> a -> a -> c

not :: 𝔹 -> 𝔹

(⩓) :: 𝔹 -> 𝔹 -> 𝔹 infixl 5

(⩔) :: 𝔹 -> 𝔹 -> 𝔹 infixl 4

cond :: 𝔹 -> a -> a -> a

stream𝐿 :: 𝐿 a -> 𝑆 a

list𝐼 :: 𝐼 a -> 𝐿 a

streamLL :: [a] -> 𝑆 a

lazyList𝐼 :: 𝐼 a -> [a]

fold𝐼 :: b -> (a -> b -> b) -> 𝐼 a -> b

foldk𝐼 :: b -> (a -> (b -> b) -> b -> b) -> 𝐼 a -> b

foldr𝐼 :: b -> (a -> b -> b) -> 𝐼 a -> b

map𝐼 :: (a -> b) -> 𝐼 a -> 𝐼 b

iter𝑆 :: 𝑆 a -> 𝐼 a

data Constraint

The kind of constraints, like Show a

otherwise :: Bool

otherwise is defined as the value True. It helps to make guards more readable. eg.

 f x | x < 0     = ...
     | otherwise = ...

($) :: (a -> b) -> a -> b infixr 0

Application operator. This operator is redundant, since ordinary application (f x) means the same as (f $ x). However, $ has low, right-associative binding precedence, so it sometimes allows parentheses to be omitted; for example:

f $ g $ h x  =  f (g (h x))

It is also useful in higher-order situations, such as map ($ 0) xs, or zipWith ($) fs xs.

Note that ($) is levity-polymorphic in its result type, so that foo $ True where foo :: Bool -> Int# is well-typed

class Eq a where

The Eq class defines equality (==) and inequality (/=). All the basic datatypes exported by the Prelude are instances of Eq, and Eq may be derived for any datatype whose constituents are also instances of Eq.

The Haskell Report defines no laws for Eq. However, == is customarily expected to implement an equivalence relationship where two values comparing equal are indistinguishable by "public" functions, with a "public" function being one not allowing to see implementation details. For example, for a type representing non-normalised natural numbers modulo 100, a "public" function doesn't make the difference between 1 and 201. It is expected to have the following properties:

Reflexivity

x == x = True

Symmetry

x == y = y == x

Transitivity

if x == y && y == z = True, then x == z = True

Substitutivity

if x == y = True and f is a "public" function whose return type is an instance of Eq, then f x == f y = True

Negation

x /= y = not (x == y)

Minimal complete definition: either == or /=.

Minimal complete definition

(==) | (/=)

Methods

(==) :: a -> a -> Bool infix 4

class Eq a => Ord a where

The Ord class is used for totally ordered datatypes.

Instances of Ord can be derived for any user-defined datatype whose constituent types are in Ord. The declared order of the constructors in the data declaration determines the ordering in derived Ord instances. The Ordering datatype allows a single comparison to determine the precise ordering of two objects.

The Haskell Report defines no laws for Ord. However, <= is customarily expected to implement a non-strict partial order and have the following properties:

Transitivity

if x <= y && y <= z = True, then x <= z = True

Reflexivity

x <= x = True

Antisymmetry

if x <= y && y <= x = True, then x == y = True

Note that the following operator interactions are expected to hold:

x >= y = y <= x

x < y = x <= y && x /= y

x > y = y < x

x < y = compare x y == LT

x > y = compare x y == GT

x == y = compare x y == EQ

min x y == if x <= y then x else y = True

max x y == if x >= y then x else y = True

Minimal complete definition: either compare or <=. Using compare can be more efficient for complex types.

Minimal complete definition

compare | (<=)

Methods

compare :: a -> a -> Ordering

class Show a where

Conversion of values to readable Strings.

Derived instances of Show have the following properties, which are compatible with derived instances of Read:

The result of show is a syntactically correct Haskell expression containing only constants, given the fixity declarations in force at the point where the type is declared. It contains only the constructor names defined in the data type, parentheses, and spaces. When labelled constructor fields are used, braces, commas, field names, and equal signs are also used.

If the constructor is defined to be an infix operator, then showsPrec will produce infix applications of the constructor.

the representation will be enclosed in parentheses if the precedence of the top-level constructor in x is less than d (associativity is ignored). Thus, if d is 0 then the result is never surrounded in parentheses; if d is 11 it is always surrounded in parentheses, unless it is an atomic expression.

If the constructor is defined using record syntax, then show will produce the record-syntax form, with the fields given in the same order as the original declaration.

For example, given the declarations

infixr 5 :^:
data Tree a =  Leaf a  |  Tree a :^: Tree a

the derived instance of Show is equivalent to

instance (Show a) => Show (Tree a) where

       showsPrec d (Leaf m) = showParen (d > app_prec) $
            showString "Leaf " . showsPrec (app_prec+1) m
         where app_prec = 10

       showsPrec d (u :^: v) = showParen (d > up_prec) $
            showsPrec (up_prec+1) u .
            showString " :^: "      .
            showsPrec (up_prec+1) v
         where up_prec = 5

Note that right-associativity of :^: is ignored. For example,

show (Leaf 1 :^: Leaf 2 :^: Leaf 3) produces the string "Leaf 1 :^: (Leaf 2 :^: Leaf 3)".

Minimal complete definition

showsPrec | show

Methods

show :: a -> String

A specialised variant of showsPrec, using precedence context zero, and returning an ordinary String.

data Bool

Constructors

False

True

data Ordering

Constructors

LT

EQ

GT

data IO a

A value of type IO a is a computation which, when performed, does some I/O before returning a value of type a.

There is really only one way to "perform" an I/O action: bind it to Main.main in your program. When your program is run, the I/O will be performed. It isn't possible to perform I/O from an arbitrary function, unless that function is itself in the IO monad and called at some point, directly or indirectly, from Main.main.

IO is a monad, so IO actions can be combined using either the do-notation or the >> and >>= operations from the Monad class.

undefined :: HasCallStack => a

A special case of error. It is expected that compilers will recognize this and insert error messages which are more appropriate to the context in which undefined appears.