UVMHS.Lib.Fuzzy

Documentation

class Fuzzy a where

Methods

fuzzy :: FuzzyM a

data FuzzyEnv

Constructors

FuzzyEnv

Fields

fuzzyEnvRadius :: ℕ64

fuzzyEnvDepth :: ℕ64

newtype FuzzyM a

Constructors

FuzzyM

Fields

unRandM :: RWS FuzzyEnv () RG a

fuzzyEnvRadiusL :: FuzzyEnv ⟢ ℕ64

fuzzyEnvDepthL :: FuzzyEnv ⟢ ℕ64

mkFuzzyM :: (FuzzyEnv -> RG -> RG ∧ a) -> FuzzyM a

runFuzzyM :: FuzzyEnv -> RG -> FuzzyM a -> RG ∧ a

runFuzzyMRG :: FuzzyEnv -> FuzzyM a -> State RG a

fuzzyDec :: FuzzyM a -> FuzzyM a

Use this to ensure termination when building recursive datatypes. Note that this will underflow to the maximum natural when d is zero, so you should only every use this when you're sure d is not zero, e.g. guarded behind a wrchoose with coefficient d.

fuzzyRec :: Fuzzy a => FuzzyM a

rand :: forall a m. (MonadRand m, Fuzzy a) => ℕ64 -> ℕ64 -> m a

randTny :: forall a m. (MonadRand m, Fuzzy a) => m a

randSml :: forall a m. (MonadRand m, Fuzzy a) => m a

randMed :: forall a m. (MonadRand m, Fuzzy a) => m a

randLrg :: forall a m. (MonadRand m, Fuzzy a) => m a

fuzzyRadius :: FuzzyM ℕ64

fuzzyDepth :: FuzzyM ℕ64

isoFuzzy :: (a ⇄ b, Fuzzy b) => FuzzyM a