UVMHS.Lib.Options

Documentation

option :: 𝐿 ℂ -> 𝐿 𝕊 -> ArgDescr a -> 𝕊 -> OptDescr a

noArg :: a -> ArgDescr a

reqArg :: (𝕊 -> a) -> 𝕊 -> ArgDescr a

optArg :: (𝑂 𝕊 -> a) -> 𝕊 -> ArgDescr a

optUsageInfo :: 𝕊 -> 𝐿 (OptDescr a) -> 𝕊

parseOptions :: 𝐿 (OptDescr a) -> 𝐿 𝕊 -> (𝐿 a ∧ 𝐿 𝕊) ∧ 𝐿 𝕊

data OptDescr a

Each OptDescr describes a single option.

The arguments to Option are:

list of short option characters

list of long option strings (without "--")

argument descriptor

explanation of option for user

data ArgDescr a

Describes whether an option takes an argument or not, and if so how the argument is injected into a value of type a.