UVMHS.Lib.Parser.Regex

Documentation

class Classified c t | t -> c where

Methods

classify :: t -> c

data RegexResult o u

Constructors

RegexResult

Fields

regexResultLevel :: ℕ64

regexResultFormats :: Formats

regexResultOutput :: 𝑂 o

regexResultUpdate :: u

newtype RegexInfo o u

Constructors

RegexInfo

Fields

regexInfoResult :: 𝑂 (RegexResult o u)

type Regex c t o u = Annotated (RegexInfo o u) (RegexU c t o u)

data RegexU c t o u

Constructors

NullR

ResR (RegexResult o u)

AtomR (RegexResult o u) (RegexAtom c t o u)

SumsR (𝑃 (Regex c t o u))

SeqsR (𝐿 (Regex c t o u))

StarR (RegexResult o u) (Regex c t o u)

data RegexAtom c t o u

Constructors

TokRA t

NTokRA (𝑃 t)

ClassRA c

nullRegex :: Zero u => Regex c t o u

resRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => RegexResult o u -> Regex c t o u

epsRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => Regex c t o u

retRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => ℕ64 -> Formats -> 𝑂 o -> u -> Regex c t o u

outRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => ℕ64 -> Formats -> o -> Regex c t o u

lepsRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => ℕ64 -> Regex c t o u

fepsRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => Formats -> Regex c t o u

oepsRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => o -> Regex c t o u

uepsRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => u -> Regex c t o u

atomRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => RegexAtom c t o u -> Regex c t o u

tokRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => t -> Regex c t o u

ntokRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => 𝑃 t -> Regex c t o u

classRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => c -> Regex c t o u

consEpsRegex :: (Ord c, Ord t, Ord o, Ord u, Plus u) => RegexResult o u -> Regex c t o u -> Regex c t o u

consEpsRegexU :: (Ord c, Ord t, Ord o, Ord u, Plus u) => RegexResult o u -> RegexU c t o u -> RegexU c t o u

snocEpsRegex :: (Ord c, Ord t, Ord o, Ord u, Plus u) => RegexResult o u -> Regex c t o u -> Regex c t o u

snocEpsRegexU :: (Ord c, Ord t, Ord o, Ord u, Plus u) => RegexResult o u -> RegexU c t o u -> RegexU c t o u

sumRegex :: (Ord c, Ord t, Ord o, Ord u, Plus u) => Regex c t o u -> Regex c t o u -> Regex c t o u

seqRegex :: (Ord c, Ord t, Ord o, Ord u, Additive u) => Regex c t o u -> Regex c t o u -> Regex c t o u

starRegex :: (Ord c, Ord t, Ord o, Ord u, Zero u) => Regex c t o u -> Regex c t o u

derRegex :: (Ord c, Ord t, Classified c t, Ord o, Ord u, Additive u) => (t ∨ c) -> Regex c t o u -> Regex c t o u

derRegexAtom :: (Ord c, Ord t, Classified c t, Ord o, Ord u, Additive u) => (t ∨ c) -> RegexAtom c t o u -> Regex c t o u

derRegexSequence :: (Ord t, Ord c, Classified c t, Ord o, Ord u, Additive u) => (t ∨ c) -> 𝐿 (Regex c t o u) -> Regex c t o u

regexLits :: Ord t => Regex c t o u -> 𝑃 t

regexLitsAtom :: Ord t => RegexAtom c t o u -> 𝑃 t

data RegexState c t o u

Constructors

RegexState

Fields

regexStateNextID :: ℕ64

regexStateMap :: Regex c t o u ⇰ ℕ64

regexStateTransitions :: (t ∨ c) ⇰ (ℕ64 ⇰ ℕ64)

regexStateResults :: ℕ64 ⇰ 𝑂 (RegexResult o u)

regexStateDead :: ℕ64 ⇰ 𝔹

regexStateDeadL :: forall (c :: Type) (t :: Type) (o :: Type) (u :: Type). (⟢) (RegexState c t o u) ((⇰) ℕ64 𝔹)

regexStateResultsL :: forall (c :: Type) (t :: Type) (o :: Type) (u :: Type). (⟢) (RegexState c t o u) ((⇰) ℕ64 (𝑂 (RegexResult o u)))

regexStateTransitionsL :: forall (c :: Type) (t :: Type) (o :: Type) (u :: Type). (⟢) (RegexState c t o u) ((⇰) ((∨) t c) ((⇰) ℕ64 ℕ64))

regexStateMapL :: forall (c :: Type) (t :: Type) (o :: Type) (u :: Type). (⟢) (RegexState c t o u) ((⇰) (Regex c t o u) ℕ64)

regexStateNextIDL :: forall (c :: Type) (t :: Type) (o :: Type) (u :: Type). (⟢) (RegexState c t o u) ℕ64

regexState₀ :: RegexState c t o u

data DFA c t o u

Constructors

DFA

Fields

dfaLits :: 𝑃 t

dfaStart :: ℕ64

dfaTransitions :: (t ∨ c) ⇰ 𝕍 ℕ64

dfaSuccess :: 𝕍 (𝑂 (RegexResult o u))

dfaDead :: 𝕍 𝔹

compileRegex :: forall c t o u. (Pretty c, Pretty t, Pretty o, Pretty u, Ord c, Ord t, Classified c t, All c, Ord o, Ord u, Additive u) => Regex c t o u -> DFA c t o u

data LexDFAState t

Constructors

LexDFAState

Fields

lexDFAStatePrefix :: WindowR Doc Doc

lexDFAStateContext :: ParserContext

lexDFAStateInput :: ParserInput t

lexDFAStateTokens :: 𝐼S t

data Lexer c t o u w

Constructors

Lexer

Fields

lexerDFA :: u -> DFA c t o u

lexerMkToken :: 𝐼S t -> 𝑂 o -> 𝔹 ∧ w

lexerInitState :: u

tokenize :: forall c t o u w. (Show u, Ord c, Ord t, Pretty t, Classified c t, Eq o, Eq u, Plus u) => Lexer c t o u w -> 𝕊 -> 𝕍 (ParserToken t) -> Doc ∨ 𝕍 (ParserToken w)

tokenizeIO :: forall c t o u w. (Show u, Ord c, Ord t, Pretty t, Classified c t, Eq o, Eq u, Plus u) => Lexer c t o u w -> 𝕊 -> 𝕍 (ParserToken t) -> IO (𝕍 (ParserToken w))

tokenizeIOMain :: forall c t o u w. (Show u, Ord c, Ord t, Pretty t, Classified c t, Eq o, Eq u, Plus u, Pretty w) => Lexer c t o u w -> 𝕊 -> 𝕍 (ParserToken t) -> IO ()

data CharClass

Constructors

LetterClass

NumberClass

SpaceClass

OtherClass

lWord :: (Zero u, Ord o, Ord u, Additive u) => 𝕊 -> Regex CharClass ℂ o u

lSpace :: (Zero u, Ord o, Ord u, Additive u) => Regex CharClass ℂ o u

lName :: (Zero u, Ord u, Ord o, Additive u) => Regex CharClass ℂ o u

lNatPre :: (Zero u, Ord u, Ord o, Additive u) => Regex CharClass ℂ o u

lNat :: (Zero u, Ord u, Ord o, Additive u) => Regex CharClass ℂ o u

lNatCoded :: (Zero u, Ord u, Ord o, Additive u) => Regex CharClass ℂ o u

lIntPre :: (Zero u, Ord o, Ord u, Additive u) => Regex CharClass ℂ o u

lInt :: (Zero u, Ord o, Ord u, Additive u) => Regex CharClass ℂ o u

lDbl :: (Zero u, Ord o, Ord u, Additive u) => Regex CharClass ℂ o u

lString :: (Zero u, Ord o, Ord u, Additive u) => Regex CharClass ℂ o u

lComment :: Ord o => Regex CharClass ℂ o ℕ64

lCommentMLOpen :: Ord o => Regex CharClass ℂ o ℕ64

lCommentMLBodyOpen :: Ord o => Regex CharClass ℂ o ℕ64

lCommentMLBodyClose :: Ord o => Regex CharClass ℂ o ℕ64

lCommentMLBody :: Ord o => Regex CharClass ℂ o ℕ64

data TokenClassBasic

Constructors

SpaceCBasic

CommentCBasic

SyntaxCBasic

StringCBasic

NameCBasic

NaturalCBasic

IntegerCBasic

DoubleCBasic

doubleCBasicL :: (⌲) TokenClassBasic ()

integerCBasicL :: (⌲) TokenClassBasic ()

naturalCBasicL :: (⌲) TokenClassBasic ()

nameCBasicL :: (⌲) TokenClassBasic ()

stringCBasicL :: (⌲) TokenClassBasic ()

syntaxCBasicL :: (⌲) TokenClassBasic ()

commentCBasicL :: (⌲) TokenClassBasic ()

spaceCBasicL :: (⌲) TokenClassBasic ()

data TokenBasic

Constructors

SpaceTBasic 𝕊

CommentTBasic 𝕊

SyntaxTBasic 𝕊

StringTBasic 𝕊

NameTBasic 𝕊

NaturalTBasic ℕ

IntegerTBasic ℤ

DoubleTBasic 𝔻

doubleTBasicL :: (⌲) TokenBasic 𝔻

integerTBasicL :: (⌲) TokenBasic ℤ

naturalTBasicL :: (⌲) TokenBasic ℕ

nameTBasicL :: (⌲) TokenBasic 𝕊

stringTBasicL :: (⌲) TokenBasic 𝕊

syntaxTBasicL :: (⌲) TokenBasic 𝕊

commentTBasicL :: (⌲) TokenBasic 𝕊

spaceTBasicL :: (⌲) TokenBasic 𝕊

mkTokenBasic :: 𝐼S ℂ -> 𝑂 TokenClassBasic -> 𝔹 ∧ TokenBasic

lSyntaxBasic :: (Ord u, Additive u) => 𝐿 𝕊 -> 𝐿 𝕊 -> 𝐿 𝕊 -> 𝐿 𝕊 -> Regex CharClass ℂ TokenClassBasic u

lTokenBasic :: 𝐿 𝕊 -> 𝐿 𝕊 -> 𝐿 𝕊 -> 𝐿 𝕊 -> Regex CharClass ℂ TokenClassBasic ℕ64

lCommentMLBasic :: Regex CharClass ℂ TokenClassBasic ℕ64

dfaBasic :: 𝐿 𝕊 -> 𝐿 𝕊 -> 𝐿 𝕊 -> 𝐿 𝕊 -> ℕ64 -> DFA CharClass ℂ TokenClassBasic ℕ64

lexerBasic :: 𝐿 𝕊 -> 𝐿 𝕊 -> 𝐿 𝕊 -> 𝐿 𝕊 -> Lexer CharClass ℂ TokenClassBasic ℕ64 TokenBasic