UVMHS.Lib.Substitution

Names

data Name

Constructors

Name

Fields

nameMark :: 𝑂 ℕ64

nameName :: 𝕊

nameMarkL :: Name ⟢ 𝑂 ℕ64

nameNameL :: Name ⟢ 𝕊

mkName :: 𝕊 -> Name

gensymName :: (Monad m, MonadState s m) => (s ⟢ ℕ64) -> 𝕊 -> m Name

syntaxName :: LexerBasicSyntax

pName :: CParser TokenBasic Name

pNameWS :: CParser TokenWSBasic Name

Individual Variable Types

newtype DVar

Constructors

DVar

Fields

unDVar :: ℕ64

unDVarL :: DVar ⟢ ℕ64

syntaxDVar :: LexerBasicSyntax

pDVarTail :: CParser TokenBasic DVar

pDVar :: CParser TokenBasic DVar

data NVar

Constructors

NVar

Fields

nvarIndex :: DVar

nvarName :: Name

nvarIndexL :: NVar ⟢ DVar

nvarNameL :: NVar ⟢ Name

nvar_Name :: Name -> NVar

name_NVarL :: NVar ⌲ Name

gensymNVar :: (Monad m, MonadState s m) => (s ⟢ ℕ64) -> 𝕊 -> m NVar

syntaxNVar :: LexerBasicSyntax

pNVarTail :: Name -> CParser TokenBasic NVar

pNVar :: CParser TokenBasic NVar

newtype GVar

Constructors

GVar

Fields

unGVar :: Name

unGVarL :: GVar ⟢ Name

syntaxGVar :: LexerBasicSyntax

pGVarTail :: Name -> CParser TokenBasic GVar

pGVar :: CParser TokenBasic GVar

data MVar s e

Constructors

MVar

Fields

mvarSubst :: Subst s e

mvarName :: Name

mvarSubstL :: MVar s e ⟢ Subst s e

mvarNameL :: MVar s e ⟢ Name

wfMVar :: Ord s => MVar s e -> 𝔹

canonMVar :: (Ord s, Eq e, Substy s e e) => (e -> e) -> MVar s e -> MVar s e

substMVar :: (Ord s, Substy s e e) => Subst s e -> MVar s e -> MVar s e

syntaxMVar :: LexerBasicSyntax

pMVarTail :: (Eq e, Substy () e e) => (() -> CParser TokenBasic e) -> Name -> CParser TokenBasic (MVar () e)

pMVar :: (Eq e, Substy () e e) => (() -> CParser TokenBasic e) -> CParser TokenBasic (MVar () e)

Combined Variable Types

data SVar

Constructors

D_SVar DVar

N_SVar NVar

d_SVarL :: SVar ⌲ DVar

n_SVarL :: SVar ⌲ NVar

mkSVar :: SName -> DVar -> SVar

svarName :: SVar -> SName

svarLevel :: SVar -> DVar

data Var

Constructors

D_Var DVar

N_Var NVar

G_Var GVar

d_VarL :: Var ⌲ DVar

n_VarL :: Var ⌲ NVar

g_VarL :: Var ⌲ GVar

var_Name :: Name -> Var

name_VarL :: Var ⌲ Name

gensymVar :: (Monad m, MonadState s m) => (s ⟢ ℕ64) -> 𝕊 -> m Var

syntaxVar :: LexerBasicSyntax

pVar :: CParser TokenBasic Var

data UVar s e

Constructors

D_UVar DVar

N_UVar NVar

G_UVar GVar

M_UVar (MVar s e)

d_UVarL :: UVar s e ⌲ DVar

n_UVarL :: UVar s e ⌲ NVar

g_UVarL :: UVar s e ⌲ GVar

m_UVarL :: UVar s e ⌲ MVar s e

name_UVarL :: UVar s e ⌲ Name

uvar_Name :: Name -> UVar s e

uvar_SVar :: SVar -> UVar s e

svar_UVarL :: UVar s e ⌲ SVar

uvar_Var :: Var -> UVar s e

var_UVarL :: UVar s e ⌲ Var

gensymUVar :: (Monad m, MonadState s m) => (s ⟢ ℕ64) -> 𝕊 -> m (UVar s e)

wfUVar :: Ord s => UVar s e -> 𝔹

canonUVar :: (Ord s, Eq e, Substy s e e) => (e -> e) -> UVar s e -> UVar s e

syntaxUVar :: LexerBasicSyntax

pUVar :: (Eq e, Substy () e e) => (() -> CParser TokenBasic e) -> CParser TokenBasic (UVar () e)

Variable Lens Classes

class SVarView s e | e -> s where

Methods

svarL :: s -> e ⌲ SVar

svarScopeL :: SVarView s e => s -> SName -> e ⌲ DVar

De Bruijn, Named and Global and Substitutions

data Subst s e

isNullSubst :: Ord s => Subst s e -> 𝔹

wfSubst :: Ord s => Subst s e -> 𝔹

canonSubst :: (Ord s, Eq e, Substy s e e) => (e -> e) -> Subst s e -> Subst s e

syntaxSubst :: LexerBasicSyntax

pSubst :: (Eq e, Substy () e e) => (() -> CParser TokenBasic e) -> CParser TokenBasic (Subst () e)

dshiftsSubst :: Ord s => (s ⇰ ℕ64) -> Subst s e -> Subst s e

dshiftSubst :: Ord s => s -> ℕ64 -> Subst s e -> Subst s e

dintrosSubst :: Ord s => (s ⇰ ℕ64) -> Subst s e

dintroSubst :: Ord s => s -> ℕ64 -> Subst s e

dbindsSubst :: Ord s => (s ⇰ 𝕍 e) -> Subst s e

dbindSubst :: Ord s => s -> e -> Subst s e

nshiftsSubst :: Ord s => (s ⇰ (Name ⇰ ℕ64)) -> Subst s e -> Subst s e

nshiftSubst :: Ord s => s -> Name -> ℕ64 -> Subst s e -> Subst s e

nintrosSubst :: Ord s => (s ⇰ (Name ⇰ ℕ64)) -> Subst s e

nintroSubst :: Ord s => s -> Name -> ℕ64 -> Subst s e

nbindsSubst :: Ord s => (s ⇰ (Name ⇰ 𝕍 e)) -> Subst s e

nbindSubst :: Ord s => s -> Name -> e -> Subst s e

gbindsSubst :: Ord s => (s ⇰ (Name ⇰ e)) -> Subst s e

gbindSubst :: Ord s => s -> Name -> e -> Subst s e

Metavar Substitutions

data MetaSubst s e

mbindsSubst :: Ord s => (s ⇰ (Name ⇰ e)) -> MetaSubst s e

mbindSubst :: Ord s => s -> Name -> e -> MetaSubst s e

Substy Interface

data SubstyM s e a

class SVarView s e => Substy s e a | a -> e, e -> s where

Methods

substy :: a -> SubstyM s e a

substyDBdr :: (Ord s, Ord e) => s -> SubstyM s e ()

substyNBdr :: (Ord s, Ord e) => s -> Name -> SubstyM s e ()

substyBdr :: (Ord s, Ord e, Substy s e e) => s -> (UVar s e -> e) -> Name -> SubstyM s e ()

substyDVar :: (Ord s, Ord e, Substy s e e) => (DVar -> e) -> s -> DVar -> SubstyM s e e

substyNVar :: (Ord s, Ord e, Substy s e e) => (NVar -> e) -> s -> NVar -> SubstyM s e e

substyGVar :: (Ord s, Ord e, Substy s e e) => (GVar -> e) -> s -> GVar -> SubstyM s e e

substySVar :: (Ord s, Ord e, Substy s e e) => (SVar -> e) -> s -> SVar -> SubstyM s e e

substyVar :: (Ord s, Ord e, Pretty e, Pretty s, Substy s e e) => (Var -> e) -> s -> Var -> SubstyM s e e

substyMVar :: (Ord s, Ord e, Pretty e, Pretty s, Substy s e e) => (MVar s e -> e) -> s -> MVar s e -> SubstyM s e e

substyUVar :: (Ord s, Ord e, Pretty e, Pretty s, Substy s e e) => (UVar s e -> e) -> s -> UVar s e -> SubstyM s e e

fvssWith :: Substy s e a => (s -> UVar s e -> 𝔹) -> a -> s ⇰ 𝑃 (UVar s e)

fvsWith :: (Ord s, Substy s e a) => s -> (UVar s e -> 𝔹) -> a -> 𝑃 (UVar s e)

fvss :: Substy s e a => a -> s ⇰ 𝑃 (UVar s e)

fvs :: (Ord s, Substy s e a) => s -> a -> 𝑃 (UVar s e)

fvssMetas :: (Ord s, Ord e, Substy s e a) => a -> s ⇰ 𝑃 (MVar s e)

fvsMetas :: (Ord s, Ord e, Substy s e a) => s -> a -> 𝑃 (MVar s e)

todbr :: Substy s e a => a -> 𝑂 a

tonmd :: Substy s e a => a -> 𝑂 a

subst :: Substy s e a => Subst s e -> a -> 𝑂 a

msubst :: Substy s e a => MetaSubst s e -> a -> 𝑂 a