UVMHS.Lib.Substitution.SubstScoped

Documentation

data SubstScoped s e

Constructors

SubstScoped

Fields

substScopedShift :: ℕ64

ρ: De Bruijn indices lower than this number will be untouched by this substitution. Think of it as a substitution working over all natural numbers being shifted to the right to ignore this many first indices.

substScopeElems :: 𝕍 (SSubstElem s e)

es: Instantiates as many of the first indices (post-shift by ρ) as the length of this vector with the values in the vector.

substScopeIntro :: ℤ64

ι: Starting at the nameless variable index after all the shifts (ρ) and all the instantiations (es), simulate an introduction of this many new nameless variables by bumping all subsequent indices by this much.

substScopedShiftL :: SubstScoped s e ⟢ ℕ64

substScopeElemsL :: SubstScoped s e ⟢ 𝕍 (SSubstElem s e)

substScopeIntroL :: SubstScoped s e ⟢ ℤ64

lookupSubstScoped :: SubstScoped s e -> DVar -> SSubstElem s e

interpSubstScoped :: (e ⌲ DVar) -> ((s ⇰ ℕ64) -> e -> 𝑂 e) -> SubstScoped s e -> DVar -> 𝑂 e

wfSubstScoped :: SubstScoped s e -> 𝔹

canonSubstScoped :: (Eq s, Eq e) => (e ⌲ DVar) -> ((s ⇰ ℕ64) -> e -> 𝑂 e) -> (e -> e) -> SubstScoped s e -> SubstScoped s e

isNullSubstScoped :: SubstScoped s e -> 𝔹

introSubstScoped :: ℕ64 -> SubstScoped s e

shiftSubstScoped :: Ord s => (s ⇰ ℕ64) -> s -> SubstScoped s e -> SubstScoped s e

bindSubstScoped :: 𝕍 e -> SubstScoped s e

substSubstScoped :: (e ⌲ DVar) -> ((s ⇰ ℕ64) -> e -> 𝑂 e) -> SubstScoped s e -> SubstScoped s e

ppSubstScopedWith :: Pretty e => ((s ⇰ ℕ64) -> Doc) -> (DVarInf -> Doc) -> SubstScoped s e -> 𝐼 (Doc ∧ Doc)