UVMHS.Lib.Substitution.Substy

Documentation

data FreeVarsAction s e

Constructors

FreeVarsAction

Fields

freeVarsActionFilter :: s -> 𝕐 s e -> 𝔹

freeVarsActionScope :: (s ∧ 𝑂 𝕎) ⇰ ℕ64

freeVarsActionFilterL :: FreeVarsAction s e ⟢ (s -> 𝕐 s e -> 𝔹)

freeVarsActionScopeL :: FreeVarsAction s e ⟢ ((s ∧ 𝑂 𝕎) ⇰ ℕ64)

data RebindAction

Constructors

ID_RA

AllNameless_RA

AllNamed_RA

data SubstAction s e

Constructors

SubstAction

Fields

substActionRebind :: RebindAction

substActionSubst :: Subst s e

substActionRebindL :: SubstAction s e ⟢ RebindAction

substActionSubstL :: SubstAction s e ⟢ Subst s e

data SubstyAction s e

Constructors

FreeVars_SA (FreeVarsAction s e)

Subst_SA (SubstAction s e)

MetaSubst_SA (MetaSubst s e)

freeVars_SAL :: SubstyAction s e ⌲ FreeVarsAction s e

subst_SAL :: SubstyAction s e ⌲ SubstAction s e

metaSubst_SAL :: SubstyAction s e ⌲ MetaSubst s e

newtype SubstyM s e a

Constructors

SubstyM

Fields

unSubstyM :: UContT (ReaderT (SubstyAction s e) (FailT (WriterT (s ⇰ 𝑃 (𝕐 s e)) ID))) a

mkSubstM :: (forall u. SubstyAction s e -> (a -> SubstyAction s e -> (s ⇰ 𝑃 (𝕐 s e)) ∧ 𝑂 u) -> (s ⇰ 𝑃 (𝕐 s e)) ∧ 𝑂 u) -> SubstyM s e a

runSubstM :: SubstyAction s e -> (a -> SubstyAction s e -> (s ⇰ 𝑃 (𝕐 s e)) ∧ 𝑂 u) -> SubstyM s e a -> (s ⇰ 𝑃 (𝕐 s e)) ∧ 𝑂 u

evalSubstM :: SubstyAction s e -> SubstyM s e a -> (s ⇰ 𝑃 (𝕐 s e)) ∧ 𝑂 a

class SVarView s e => Substy s e a | a -> e, e -> s where

Methods

substy :: a -> SubstyM s e a

fvsWith :: Substy s e a => (s -> 𝕐 s e -> 𝔹) -> a -> s ⇰ 𝑃 (𝕐 s e)

fvsSMetas :: (Ord s, Ord e, Substy s e a) => 𝑃 s -> a -> s ⇰ 𝑃 (𝕎 ∧ Subst s e)

fvsMetas :: (Ord s, Ord e, Substy s e a) => s -> a -> 𝑃 (𝕎 ∧ Subst s e)

fvs :: Substy s e a => a -> s ⇰ 𝑃 (𝕐 s e)

todbr :: Substy s e a => a -> 𝑂 a

tonmd :: Substy s e a => a -> 𝑂 a

subst :: Substy s e a => Subst s e -> a -> 𝑂 a

msubst :: Substy s e a => MetaSubst s e -> a -> 𝑂 a

canonSubst :: (Ord s, Eq e, Substy s e e) => (e -> e) -> Subst s e -> Subst s e

canonUVar :: (Ord s, Eq e, Substy s e e) => (e -> e) -> 𝕐 s e -> 𝕐 s e

nullSubst :: Subst s e

appendSubst :: (Ord s, Substy s e e) => Subst s e -> Subst s e -> Subst s e

substyDBdr :: (Ord s, Ord e) => s -> SubstyM s e ()

substyNBdr :: (Ord s, Ord e) => s -> 𝕎 -> SubstyM s e ()

substyBdr :: (Ord s, Ord e, Substy s e e) => s -> (𝕐 s e' -> e) -> 𝕎 -> SubstyM s e ()

substyVar :: (Ord s, Ord e, Substy s e e) => 𝑂 𝕎 -> s -> (ℕ64 -> e) -> ℕ64 -> SubstyM s e e

substyDVar :: (Ord s, Ord e, Substy s e e) => s -> (ℕ64 -> e) -> ℕ64 -> SubstyM s e e

substyNVar :: (Ord s, Ord e, Substy s e e) => s -> (ℕ64 -> e) -> 𝕎 -> ℕ64 -> SubstyM s e e

substyGVar :: (Ord s, Ord e, Substy s e e) => s -> (𝕎 -> e) -> 𝕎 -> SubstyM s e e

substyMVar :: (Ord s, Ord e, Pretty e, Pretty s, Substy s e e) => s -> (𝕎 -> Subst s e -> e) -> 𝕎 -> Subst s e -> SubstyM s e e

substy𝕏 :: (Ord s, Ord e, Pretty e, Pretty s, Substy s e e) => s -> (𝕏 -> e) -> 𝕏 -> SubstyM s e e

substy𝕐 :: (Ord s, Ord e, Pretty e, Pretty s, Substy s e e) => s -> (𝕐 s e -> e) -> 𝕐 s e -> SubstyM s e e

syntaxUVar :: LexerBasicSyntax

data ParseSubstAction e

Constructors

ParseSubstAction

Fields

parseSubstActionShfts :: 𝐼 ℕ64

parseSubstActionElems :: 𝑂 ℕ64 ⇰ 𝐼 e

parseSubstActionIncrs :: 𝐼 (ℕ64 ∧ ℤ64)

parseSubstActionShftsL :: ParseSubstAction e ⟢ 𝐼 ℕ64

parseSubstActionElemsL :: ParseSubstAction e ⟢ (𝑂 ℕ64 ⇰ 𝐼 e)

parseSubstActionIncrsL :: ParseSubstAction e ⟢ 𝐼 (ℕ64 ∧ ℤ64)

parseSubstActionShft :: ℕ64 -> ParseSubstAction e

parseSubstActionElem :: 𝑂 ℕ64 -> e -> ParseSubstAction e

parseSubstActionIncr :: ℕ64 -> ℤ64 -> ParseSubstAction e

type ParseSubstActions e = 𝑂 (𝕎 ∧ 𝔹) ⇰ ParseSubstAction e

cpSubst :: (Eq e, Substy () e e) => (() -> CParser TokenBasic e) -> CParser TokenBasic (Subst () e)

cpUVarNGMVar :: (Eq e, Substy () e e) => (() -> CParser TokenBasic e) -> CParser TokenBasic (𝕐 () e)

cpUVar :: (Eq e, Substy () e e) => (() -> CParser TokenBasic e) -> CParser TokenBasic (𝕐 () e)

cpUVarRaw :: (Eq e, Substy () e e) => (() -> CParser TokenBasic e) -> CParser TokenBasic (𝕐 () e)

Orphan instances

(Ord s, Substy s e e) => Append (Subst s e)

Instance details

Methods

(⧺) :: Subst s e -> Subst s e -> Subst s e

(Ord s, Substy s e e) => Monoid (Subst s e)

Instance details

Null (Subst s e)

Instance details

Methods

null :: Subst s e

(Ord s, Shrinky e) => Shrinky (𝕐 s e)

Instance details

Methods

shrink :: 𝕐 s e -> 𝐼 (𝕐 s e)