UVMHS.Lib.Substitution

Datatypes

Raw Variables

data 𝕎

Constructors

𝕎

Fields

varMark :: 𝑂 ℕ64

varName :: 𝕊

varMarkL :: 𝕎 ⟢ 𝑂 ℕ64

varNameL :: 𝕎 ⟢ 𝕊

var :: 𝕊 -> 𝕎

gensymVar :: (Monad m, MonadState s m) => (s ⟢ ℕ64) -> 𝕊 -> m 𝕎

data 𝕏

Constructors

D_SVar ℕ64

N_SVar ℕ64 𝕎

G_SVar 𝕎

class SVarView s e | e -> s where

Methods

svarL :: s -> e ⌲ 𝕏

data 𝕐 s e

Constructors

S_UVar 𝕏

M_UVar 𝕎 (Subst s e)

s_UVarL :: 𝕐 s e ⌲ 𝕏

data Subst s e

data MetaSubst s e

class SVarView s e => Substy s e a | a -> e, e -> s where

Methods

substy :: a -> SubstyM s e a

Variable Operations

Validating Variables

wfUVar :: Ord s => 𝕐 s e -> 𝔹

Canonicalizing Variables

canonUVar :: (Ord s, Eq e, Substy s e e) => (e -> e) -> 𝕐 s e -> 𝕐 s e

Substitution Operations

Validating Substitutions

wfSubst :: Ord s => Subst s e -> 𝔹

Canonicalizing Substitutions

canonSubst :: (Ord s, Eq e, Substy s e e) => (e -> e) -> Subst s e -> Subst s e

Building Substitutions

introDSubst :: ℕ64 -> Subst () e

Applying Substitutions

subst :: Substy s e a => Subst s e -> a -> 𝑂 a

Implementing Substy for user-defined types.

substyDBdr :: (Ord s, Ord e) => s -> SubstyM s e ()

substyBdr :: (Ord s, Ord e, Substy s e e) => s -> (𝕐 s e' -> e) -> 𝕎 -> SubstyM s e ()

substy𝕏 :: (Ord s, Ord e, Pretty e, Pretty s, Substy s e e) => s -> (𝕏 -> e) -> 𝕏 -> SubstyM s e e

substy𝕐 :: (Ord s, Ord e, Pretty e, Pretty s, Substy s e e) => s -> (𝕐 s e -> e) -> 𝕐 s e -> SubstyM s e e

Parsing

Constructing Syntax Categories

syntaxUVar :: LexerBasicSyntax

Parsing Logic

cpVar :: CParser TokenBasic 𝕎

cpUVarRaw :: (Eq e, Substy () e e) => (() -> CParser TokenBasic e) -> CParser TokenBasic (𝕐 () e)